Itemized Inventory Sheet
Example of all current products featured in both stores. Some items are sold in both stores, Decatur and Lawrenceville locations, while others are featured at one or the other. Decatur carries THC Free and Low-THC (<0.3%) but Lawrenceville can only carry THC Free products as of now. This is an issue with the county; however, these products are legal nationwide and are in compliance with the 2018 Farm Bill. Silly local government is behind the times, but eventually, all stores will carry the same products.